Extract and export the `dataEnum` serializer
