+++ packages/umi-serializer-beet/src/createBeetSerializer.ts
@@ -1,8 +1,5 @@
 import {
   BoolSerializerOptions,
-  DataEnum,
-  DataEnumSerializerOptions,
-  DataEnumToSerializerTuple,
   fixSerializer,
   mergeBytes,
   none,
@@ -14,6 +11,7 @@ import {
 } from '@metaplex-foundation/umi';
 import { array } from './array';
 import { bytes } from './bytes';
+import { dataEnum } from './dataEnum';
 import {
   BeetSerializerError,
   DeserializingEmptyBufferError,
@@ -21,7 +19,6 @@ import {
 } from './errors';
 import { getSizeDescription } from './getSizeDescription';
 import { map } from './map';
-import { maxSerializerSizes } from './maxSerializerSizes';
 import { nullable } from './nullable';
 import {
   i128,
@@ -42,7 +39,6 @@ import { publicKey } from './pubkey';
 import { scalarEnum } from './scalarEnum';
 import { set } from './set';
 import { struct } from './struct';
-import { sumSerializerSizes } from './sumSerializerSizes';
 import { tuple } from './tuple';
 
 export type BeetSerializerOptions = {
@@ -77,73 +73,6 @@ function getTolerantSerializerFactory<
 export function createBeetSerializer(
   options: BeetSerializerOptions = {}
 ): SerializerInterface {
-  const dataEnum = <T extends DataEnum, U extends T = T>(
-    variants: DataEnumToSerializerTuple<T, U>,
-    options: DataEnumSerializerOptions = {}
-  ): Serializer<T, U> => {
-    const prefix = options.size ?? u8();
-    const fieldDescriptions = variants
-      .map(
-        ([name, serializer]) =>
-          `${String(name)}${serializer ? `: ${serializer.description}` : ''}`
-      )
-      .join(', ');
-    const allVariantHaveTheSameFixedSize = variants.every(
-      (one, i, all) => one[1].fixedSize === all[0][1].fixedSize
-    );
-    const fixedVariantSize = allVariantHaveTheSameFixedSize
-      ? variants[0][1].fixedSize
-      : null;
-    const maxVariantSize = maxSerializerSizes(
-      variants.map(([, field]) => field.maxSize)
-    );
-    return {
-      description:
-        options.description ??
-        `dataEnum(${fieldDescriptions}; ${prefix.description})`,
-      fixedSize:
-        variants.length === 0
-          ? prefix.fixedSize
-          : sumSerializerSizes([prefix.fixedSize, fixedVariantSize]),
-      maxSize:
-        variants.length === 0
-          ? prefix.maxSize
-          : sumSerializerSizes([prefix.maxSize, maxVariantSize]),
-      serialize: (variant: T) => {
-        const discriminator = variants.findIndex(
-          ([key]) => variant.__kind === key
-        );
-        if (discriminator < 0) {
-          throw new BeetSerializerError(
-            `Invalid data enum variant. Got "${variant.__kind}", expected one of ` +
-              `[${variants.map(([key]) => key).join(', ')}]`
-          );
-        }
-        const variantPrefix = prefix.serialize(discriminator);
-        const variantSerializer = variants[discriminator][1];
-        const variantBytes = variantSerializer.serialize(variant as any);
-        return mergeBytes([variantPrefix, variantBytes]);
-      },
-      deserialize: (bytes: Uint8Array, offset = 0) => {
-        if (bytes.slice(offset).length === 0) {
-          throw new DeserializingEmptyBufferError('dataEnum');
-        }
-        const [discriminator, dOffset] = prefix.deserialize(bytes, offset);
-        offset = dOffset;
-        const variantField = variants[Number(discriminator)] ?? null;
-        if (!variantField) {
-          throw new BeetSerializerError(
-            `Data enum index "${discriminator}" is out of range. ` +
-              `Index should be between 0 and ${variants.length - 1}.`
-          );
-        }
-        const [variant, vOffset] = variantField[1].deserialize(bytes, offset);
-        offset = vOffset;
-        return [{ __kind: variantField[0], ...(variant ?? {}) } as U, offset];
-      },
-    };
-  };
-
   const string = (
     options: StringSerializerOptions = {}
   ): Serializer<string> => {
