+++ packages/umi-serializer-data-view/src/index.ts
@@ -1,6 +1,7 @@
 export * from './array';
 export * from './createDataViewSerializer';
 export * from './bytes';
+export * from './dataEnum';
 export * from './errors';
 export * from './map';
 export * from './nullable';
